split the commands_functions #54
