+++ src/pieces/commands/__init__.py
@@ -1,16 +1,18 @@
 from .cli_loop import loop, find_most_similar_command
 from .autocommit import git_commit
 from .commands_functions import (version,
-                                 save_asset,
-                                 edit_asset,
-                                 list_assets,
-                                 open_asset,
-                                 create_asset,
                                  ask,
                                  search,
-                                 delete_asset,
                                  change_model,
-                                 set_parser,startup,list_models,list_apps)
+                                 set_parser,startup)
+from .assets import (save_asset,
+                    edit_asset,
+                    list_assets,
+                    open_asset,
+                    create_asset,
+                    delete_asset,
+                    list_models,
+                    list_apps)
 __all__ = ['loop', 
            'find_most_similar_command',
            'git_commit',
